Add unit tests for Weather service
